feat: new '/accumulated' api consept, &insert notify selection to Shared
