+++ src/router.js
@@ -14,6 +14,7 @@ const unitRouter = require('./unit/main.js');// shift to "units"
 const unitsRouter = require('./units/main.js');
 const imgRouter = require('./img/main.js');
 
+const shareRouter = require('./share/main.js');
 const inspireRouter = require('./inspire/main.js');
 const nounsRouter = require('./nouns/main.js');
 const generalRouter = require('./general/main.js');
@@ -41,6 +42,7 @@ router.use('/units', unitsRouter)
 
 router.use('/account', accountRouter)
 
+router.use('/share', shareRouter)
 router.use('/inspire', inspireRouter)
 router.use('/nouns', nounsRouter)
 router.use('/general', generalRouter)
