+++ src/share/main.js
@@ -1,11 +1,10 @@
 const express = require('express');
 const main = express.Router();
 
-const plainExecutive = require('./plain.js');
 const accumulatedExecutive = require('./accumulated.js');
 
 main.use('/accumulated', accumulatedExecutive)
 
-main.use('/', plainExecutive)
+//main.use('/', plainExecutive)
 
 module.exports = main;
