+++ src/user/actions/main.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const main = express.Router();
 
-const sharedsExecutive = require('./shareds/execute.js');
 const threadsExecutive = require('./threads.js');
 
 main.param("id", (req, res, next, id)=>{
@@ -9,8 +8,6 @@ main.param("id", (req, res, next, id)=>{
   console.log(req.requesterId);
   next();
 })
-
-main.use('/shareds', sharedsExecutive)
 
 main.use('/threads', threadsExecutive)
